Redesign Subject detail screen and update attendance calculation type
- Change `Subject.attendancePercentage()` return type from `Double` to `Float` for better integration with Compose and Glance UI components.
- Completely redesign the `Subject` detail screen with a modern, card-based layout.
- Add a circular progress visualization for attendance that includes a visual marker for the required minimum percentage.
- Implement filtering for attendance records using `FilterChip` components (Present, Absent, Cancelled).
- Replace the legacy table-style record list with a new `AttendanceRecord.Card` component using `ElevatedCard`.
- Improve UI responsiveness by using `derivedStateOf` for record filtering and `rememberScrollState` for filter chips.
- Update `OverviewWidget` to reflect the `Float` return type change in the backend entity.

+++ app/src/main/java/com/jaguar/attendancetracker/backend/entities/Subject.kt
@@ -22,8 +22,8 @@ data class Subject(
     val attendedClasses: Int = 0,
     val minAttendance: Int = 75
 ) {
-    fun attendancePercentage(): Double = if (totalClasses == 0) 0.0
-    else (attendedClasses.toDouble() / totalClasses) * 100
+    fun attendancePercentage(): Float = if (totalClasses == 0) 0f
+    else (attendedClasses.toFloat() / totalClasses) * 100
 
     fun requiredToMakeUp(): Int {
         if (totalClasses == 0) return 0
